+++ cypress.config.js
@@ -1,0 +1,12 @@
+const { defineConfig } = require("cypress");
+
+module.exports = defineConfig({
+  e2e: {
+    setupNodeEvents(on, config) {
+      // implement node event listeners here
+    },
+  },
+  env: {
+    baseUrl: "http://localhost:5010", // TODO - Load port from src/resources/config/app.yaml
+  }
+});