+++ cypress/e2e/index.spec.cy.js
@@ -1,0 +1,11 @@
+import { Paths } from './paths'
+
+describe('Index page', () => {
+  it('Contains link to game page', () => {
+    cy.visit(Paths.baseUrl())
+    cy.get("#preposition-trainer").should("exist")
+      .click().then(() => {
+        cy.url().should('include', Paths.prepositionTrainer)
+      })
+  })
+})