+++ cypress/e2e/paths.js
@@ -1,0 +1,11 @@
+import { Env } from "./env";
+
+export const Paths = {
+  prepositionTrainer: '/preposition-trainer',
+  baseUrl: function() {
+    return Cypress.env(Env.baseUrl)
+  },
+  url: function (path) {
+    return Paths.baseUrl() + path
+  }
+}