+++ shell/run.tests.sh
@@ -2,6 +2,6 @@
 
 source ./pre_run.sh
 
-printf "\nStarting tests\n\n"
+printf "\nStarting unit tests\n\n"
 
 python3 -m unittest discover -s test -p "*test.py"