+++ src/fastergerman/game/__init__.py
@@ -2,5 +2,5 @@ from fastergerman.game.default_questions import DEFAULT_QUESTIONS
 from fastergerman.game.game import Game, Question, Score, Settings
 from fastergerman.game.game_file import delete_game, get_game_names, get_game_to_load, load_game, save_game
 from fastergerman.game.game_file import GAME_TO_LOAD_FILE_PATH, GAMES_DIR_PATH
-from fastergerman.game.game_timer import GameTimer, GameTimerState, SimpleGameTimer, TimerError
-from fastergerman.game.game_session import GameEventListener, GameSession, GameCounters
+from fastergerman.game.game_timer import AbstractGameTimer, GameTimerState, GameTimer, TimerError
+from fastergerman.game.game_session import GameEventListener, GameSession, GameTimers