+++ src/test/ui/ui_game_timer_test.py
@@ -1,7 +1,7 @@
 import unittest
 import tkinter as tk
 
-from fastergerman.game import GameTimer
+from fastergerman.game import AbstractGameTimer
 from fastergerman.ui import UIGameTimer
 from test.game import GameTimerCommon
 
@@ -10,7 +10,7 @@ class UIGameTimerTest(GameTimerCommon.GameTimerTest):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
 
-    def _create_timer(self, interval_millis: int = 1000) -> GameTimer:
+    def _create_timer(self, interval_millis: int = 1000) -> AbstractGameTimer:
         return UIGameTimer(tk.Tk(), interval_millis)